# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Personal Spending Decision Support API
  version: 1.2.0
  description: >-
    Versioned API for personal spending forecasts, unusual-spending alerts,
    and transparent recommendations. Unusual-spending results are not fraud
    classifications, and recommendations are not professional financial advice.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
...
paths:
...
  /analysis/run:
    post:
      tags: [Analysis]
      summary: Run forecasting, unusual-spending analysis, and recommendations
      x-request-example:
        use_stored_transactions: true
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                use_stored_transactions:
                  type: boolean
                  default: true
            example:
              use_stored_transactions: true
      responses:
        '200':
          description: Analysis completed and stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
              example:
                success: true
                data:
                  analysis_run_id: d5549350-6e0c-48cf-82b7-b19dadbcd119
                  generated_at: '2026-07-25T15:00:00Z'
                  forecast:
                    model_version: v1
                    period_start: '2026-07-27'
                    period_end: '2026-08-02'
                    predicted_spending: 14250.75
                    currency: KES
                  alert_summary:
                    unusual_spending_detected: true
                    alert_count: 1
                  recommendations:
                    - recommendation_code: FORECAST_EXCEEDS_BUDGET
                      title: Forecast is above budget
                      message: Forecasted spending is above the current budget.
                      severity: high
                      reason: Forecasted spending is greater than the configured budget.
                      supporting_values:
                        forecasted_spending: 14250.75
                        budget_amount: 14000
                      suggested_action: Review the largest planned categories.
                      disclaimer: Automated decision support, not professional financial advice.
        '422':
          description: Insufficient history or invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: INSUFFICIENT_HISTORY
                  message: More transaction history is required before a forecast can be produced.
  /analysis/latest:
    get:
      tags: [Analysis]
      summary: Get the most recent stored analysis
      x-request-example: {}
      responses:
        '200':
          description: Latest analysis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
              example:
                success: true
                data:
                  analysis_run_id: d5549350-6e0c-48cf-82b7-b19dadbcd119
                  generated_at: '2026-07-25T15:00:00Z'
                  forecast:
                    model_version: v1
                    period_start: '2026-07-27'
                    period_end: '2026-08-02'
                    predicted_spending: 14250.75
                    currency: KES
                  alert_summary:
                    unusual_spending_detected: true
                    alert_count: 1
                  recommendations: []
        '404':
          $ref: '#/components/responses/NotFound'
  /forecasts:
    get:
      tags: [Analysis]
      summary: List stored forecasts for the current user
      x-request-example:
        page: 1
        per_page: 20
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Forecast history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
              example:
                success: true
                data:
                  items:
                    - id: b8fa1fbe-a7d2-4ea3-929c-7aad219ef234
                      created_at: '2026-07-25T15:00:00Z'
                      model_version: v1
                      predicted_spending: 14250.75
                      currency: KES
                  page: 1
                  per_page: 20
                  total: 1
  /alerts:
    get:
      tags: [Analysis]
      summary: List stored unusual-spending alerts
      x-request-example:
        page: 1
        per_page: 20
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Alert history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
              example:
                success: true
                data:
                  items:
                    - id: 118c64f3-027a-4d20-9cf0-3496ce4e3065
                      created_at: '2026-07-25T15:00:00Z'
                      is_unusual_spending: true
                      anomaly_score: 0.032
                      explanation: Amount and recent spending change were atypical.
                  page: 1
                  per_page: 20
                  total: 1
  /recommendations:
    get:
      tags: [Analysis]
      summary: List stored recommendations
      x-request-example:
        page: 1
        per_page: 20
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200':
          description: Recommendation history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
              example:
                success: true
                data:
                  items:
                    - id: 87fa3d58-f667-4e02-84af-afab07f36cf7
                      created_at: '2026-07-25T15:00:00Z'
                      recommendation_code: FORECAST_EXCEEDS_BUDGET
                      title: Forecast is above budget
                      severity: high
                  page: 1
                  per_page: 20
                  total: 1
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PerPage:
      name: per_page
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
...
  responses:
...
    NotFound:
      description: Resource not found for the current user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: The requested resource was not found.
...
  schemas:
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            details:
              type: object
              additionalProperties: true
...
    Pagination:
      type: object
      properties:
        page: {type: integer}
        per_page: {type: integer}
        total: {type: integer}
    Recommendation:
      type: object
      required:
        - recommendation_code
        - title
        - message
        - severity
        - reason
        - supporting_values
        - suggested_action
        - disclaimer
      properties:
        recommendation_code: {type: string}
        title: {type: string}
        message: {type: string}
        severity:
          type: string
          enum: [info, positive, warning, high, critical]
        reason: {type: string}
        supporting_values:
          type: object
          additionalProperties: true
        suggested_action: {type: string}
        disclaimer: {type: string}
    AnalysisResponse:
      type: object
      properties:
        success: {type: boolean}
        data:
          type: object
          properties:
            analysis_run_id: {type: string, format: uuid}
            generated_at: {type: string, format: date-time}
            forecast:
              type: object
              properties:
                model_version: {type: string}
                period_start: {type: string, format: date}
                period_end: {type: string, format: date}
                predicted_spending: {type: number}
                currency: {type: string}
            alert_summary:
              type: object
              properties:
                unusual_spending_detected: {type: boolean}
                alert_count: {type: integer}
            recommendations:
              type: array
              items: {$ref: '#/components/schemas/Recommendation'}
    HistoryResponse:
      type: object
      properties:
        success: {type: boolean}
        data:
          allOf:
            - $ref: '#/components/schemas/Pagination'
            - type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    additionalProperties: true